requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.0.3
numpy==1.26.4
yfinance==0.2.48
ta==0.11.0
anthropic==0.34.0
orjson==3.10.7
websockets==12.0

